requirements.txt
grpcio-tools
json
redis